move taboption to components
