+++ app/packages/components/src/components/PopoutSectionTitle/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PopoutSectionTitle";
