+++ app/packages/components/src/components/TabOption/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TabOption";
