+++ app/packages/components/src/components/index.ts
@@ -4,7 +4,9 @@ export { default as Header } from "./Header";
 export * from "./Icons";
 export { default as Link } from "./Link";
 export { default as Loading } from "./Loading";
+export { default as PopoutSectionTitle } from "./PopoutSectionTitle";
 export { default as RouteRenderer } from "./RouteRenderer";
 export { default as Selector } from "./Selector";
+export { default as TabOption } from "./TabOption";
 export { default as Tooltip } from "./Tooltip";
 export { default as Theme, ThemeContext, useTheme } from "./Theme";
